Fix follow and unfollow counting

+++ Services/Friendships/FriendshipAppService.cs
@@ -36,7 +36,11 @@ public class FriendshipAppService : IFriendshipAppService
         var user = await _userManager.GetUserAsync(_httpContextAccessor.HttpContext?.User);
         var friend = await _context.Users.AsNoTracking().FirstOrDefaultAsync(e => e.Id.Equals(friendId));
 
+        if (user == null) throw new KeyNotFoundException();
         if (friend == null) throw new KeyNotFoundException();
+
+        var existing = await _context.Friendships.AsNoTracking().Where(x => x.UserId.Equals(user.Id)).Where(x => x.FriendId.Equals(friendId)).FirstOrDefaultAsync();
+        if (existing != null) throw new FormatException("Already following.");
 
         Friendship newFriend = new Friendship
         {
@@ -50,8 +54,21 @@ public class FriendshipAppService : IFriendshipAppService
         try
         {
             await _context.Friendships.AddAsync(newFriend);
-            await _context.SaveChangesAsync();
             newFriend.Friend = friend;
+
+            await _context.SaveChangesAsync();
+
+            user.FollowingCount = await GetFollowingCountAsync(user);
+            friend.FollowerCount = await GetFollowerCountAsync(friend);
+
+            _context.Users.Attach(user);
+            var updatedCurrentUser = _context.Users.Update(user).Entity;
+
+            _context.Users.Attach(friend);
+            var updatedFriend = _context.Users.Update(friend).Entity;
+
+            await _context.SaveChangesAsync();
+
             return _mapper.Map<Friendship, FriendshipDto>(newFriend);
         }
         catch (System.Exception)
@@ -63,10 +80,26 @@ public class FriendshipAppService : IFriendshipAppService
     public async Task DeleteFollow(long friendId)
     {
         var user = await _userManager.GetUserAsync(_httpContextAccessor.HttpContext?.User);
+        var friend = await _context.Users.AsNoTracking().FirstOrDefaultAsync(e => e.Id.Equals(friendId));
         var friendship = await _context.Friendships.AsNoTracking().Where(x => x.UserId.Equals(user.Id)).Where(x => x.FriendId.Equals(friendId)).SingleAsync();
+
+        if (user == null) throw new KeyNotFoundException();
+        if (friend == null) throw new KeyNotFoundException();
 
         if (friendship == null) throw new NullReferenceException();
         _context.Friendships.Remove(friendship);
+
+        await _context.SaveChangesAsync();
+
+        user.FollowingCount = await GetFollowingCountAsync(user);
+        friend.FollowerCount = await GetFollowerCountAsync(friend);
+
+        _context.Users.Attach(user);
+        var updatedCurrentUser = _context.Users.Update(user).Entity;
+
+        _context.Users.Attach(friend);
+        var updatedFriend = _context.Users.Update(friend).Entity;
+
         await _context.SaveChangesAsync();
     }
 
@@ -82,5 +115,15 @@ public class FriendshipAppService : IFriendshipAppService
         var user = await _userManager.GetUserAsync(_httpContextAccessor.HttpContext?.User);
         var followers = await _context.Friendships.AsNoTracking().Include(x => x.Friend).OrderByDescending(x => x.CreatedAt).Where(x => x.UserId == user.Id).ToListAsync();
         return _mapper.Map<List<Friendship>, List<FriendshipDto>>(followers);
+    }
+
+    public async Task<long> GetFollowerCountAsync(ApplicationUser user)
+    {
+        return await _context.Friendships.AsNoTracking().Include(x => x.User).OrderByDescending(x => x.CreatedAt).Where(x => x.FriendId == user.Id).LongCountAsync();
+    }
+
+    public async Task<long> GetFollowingCountAsync(ApplicationUser user)
+    {
+        return await _context.Friendships.AsNoTracking().Include(x => x.Friend).OrderByDescending(x => x.CreatedAt).Where(x => x.UserId == user.Id).LongCountAsync();
     }
 }
